+++ lib/src/core/design/design.dart
@@ -28,6 +28,7 @@ export 'components/hh_conditional_field.dart';
 export 'components/hh_meter.dart';
 export 'components/hh_progress.dart';
 export 'components/hh_selection.dart';
+export 'components/hh_sheet.dart';
 export 'components/hh_states.dart';
 export 'components/hh_text_field.dart';
 export 'hh_brand.dart';